+++ uninstall.php
@@ -18,6 +18,7 @@ $options = array(
 	'low_mm_style_colors',
 	'low_mm_custom_css',
 	'low_mm_panel_max_width',
+	'low_mm_button_text_align',
 	'low_mm_pq_gen',
 );